Split Spark into a separate image.
With Jupyter in the same container its easy for the
secrets for Iceberg to be revealed. In practice we
would likely want a separate Jupyter instance anyway
and this gives a chance to try Spark Connect as a way
of connecting remotely to a spark cluster.

A separate Jupyter setup will follow.