Put Traefik in front of Trino
Trino requires TLS for all connections that have authentication
enabled. This can be done by a load balancer or configuring Trino
itself. Give we will eventually put in load balancer for the
whole project we'll choose Traefik for now to achieve this.

This commit only sets up basic http forwarding and not https.