Clean up all_destroy/all_create
When creating the whole system from scratch the /etc/hosts
and $HOME/.ssh/known_hosts files are configured to avoid
the setup steps stopping and waiting for confirmation
to accept the remote key. The VMs have only just be created
through the openstack interface so this should be safe.

When destroying everything the process is essentially reversed
to avoid any manual clean up being required.