Replace Jupyter notebook server with JupyterHub
JupyterHub is intended for multi-useer environments.
The hub state and user directories are currently stored on
the node and will be wiped out on VM deletion.

An admin user is created and its password must be set the
first time the user accesses the hub.

The version of Python in the spark image has been changed
to match that running in Jupyter. The worker/driver node
versions must match