Fix Traefik dashboard router definition
The stripPrefix middleware is required so that the
X-Forwarded-Prefix header is added to the request
to the proxied service. The dashboard uses this to
construct URLs for static assets.

The HeaderRegxp matcher ensures that any api requests
are also captured as Traefik currently requests the api
at '/api/' without consulting the X-Forwarded-Prefix.
A workaround suggsted in https://github.com/traefik/traefik/issues/5853#issuecomment-1935608542