Deploy a custom build of Apache Superset on the path /analytics/superset
It includes a default connection to Trino and the
warehouse catalog.

This required reorganizing the group_vars so that we
can have separate postgress passwords for the 2 databases
deployed in iceberg_catalog and superset.

This has required significant changes to the Superset sources
that are currently under review by the project.