Use registered DNS entry and assign floating IP to Traefik
The system is now accessible on the new DNS name.

In order to assign a floating IP to Traefik directly the instance
must not be on the Internal network. We create a new virtual
network and move all nodes here. The nodes are now not directly
accessible by ssh so a new management_vm is created to act as a
SSH jump node. The inventory configures ansible to access the nodes
through the jump node.