ELT script for Opralog
It is able to extract a full copy of each of the source
tables to a set of parquet files and store in the local
filesystem or an S3 bucket.

Eventually this is likely to be an Airflow task but now it
can just be run as a cron job.