Overhaul Influx DLT scripts
Remove pyiceberg as a destination and instead
using dlt to load data to staging bucket then
kick off a Spark job to dedupliate & merge the
data in to the combined machinestate table.
Spark's SQL supports the MERGE INTO command out
of the box making the code to update the final
table easier to follow.

This also removes the source tables from the catalog
as they shouldn't be needed and can just exist as
parquet files in the object store that can be read if
necessary.