Rework Influx machinestate pipeline
It seems dlt does not deal well with schemas that have
1000s of tables in them. When a pipeline is loaded dlt
inspects all tables in a schema to find any nested tables,
which is an O(N^2) operation as it checks every table
against every other table. Our schema grew to 20000 tables
and this caused even if the `dlt pipeline influxdb info` command
to take > 5 mins to show the pipeline information.

To make things worse it does this operation on normalization
so even for channels that have very little data, e.g. 5 entries,
it would take 10 minutes to run the normalize operation.

We split the work up now by Vsystem database name and define a
separate dlt.source for each Vsystem database. This reduces the
number of tables in each schema to a few hundred and this seems
more manageable.