Configure dbt sources rather than using our ugly macro
This now assumes the extract and load process has pushed
the data into the iceberg catalog and not just the object store.
This model is much easier to work rathern with than requiring each model
load the source files.
It will also facilitate skipping the transform process if no new
records were loading all of the files again to see if there is anything
new.