fix(pipelines-common): Trim trailing slashes from pyiceberg bucket_url config parameter (#52)
### Summary

Using the `pyiceberg` destination it was possible to create a table
using a location that had empty path components, .e.g.
`s3://mybucket//dataset_name/table_name` by using a `bucket_url` with a
trailing slash (`s3://mybucket/` in this example). This would raise an
error when the files were accessed in the object storage.

This changes trims trailing slashes in the config parameter.

*There is no associated issue.*