Begin working on a Lakeeper deployment
It will replace nessie as the Iceberg catalog implementation.
Nessie has some interesting features such as the git-like model
for catalogs but this adds quite complicated extra steps in managing
it to ensure the data volumes don't grow unmanageably.

Lakeeper is a rust-based catalog, so should require fewer resources,
and also has nice features such as

  - better authentication & authorization support including token dispensing in the GUI
  - direct catalog interaction through the GUI
  - expose permissions to clients through open policy agent
  - automated table maintenance in the works