Convert dbt to use Trino over Spark
Trino is far easier to setup and doesn't suffer issues with
lack of connectivity to the driver. So far the queries
are also faster than those in Spark as the cluster
is already running so there is no overhead of the Spark
driver creation.

profiles.yml needs to contain sensitive information so has
been removed.