feat(pipelines-common): Make pyiceberg destination table_location_layout optional (#66)
The catalog will create a UUID for the path components if no location is
provided. This should discourage interaction with the S3 layer directly
as it will not be so obvious what files belong to what table.

Aside:

The e2e tests started to fail because the Bitnami minio image no longer
created the correct buckets on startup. We switch to the official minio
image and create the bucket manually.